move time-matching code from _comparison to matching
